requirements.txt
# PromptAlbumBuilder Dependencies

# Image processing
pillow>=10.0.0

# EXIF metadata extraction
exifread>=3.0.0

# Configuration file parsing
pyyaml>=6.0.0

# Web UI
streamlit>=1.28.0

# Testing frameworks
hypothesis>=6.0.0
pytest>=7.0.0
